mspr3-deployment/shared/docker/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
psycopg2-binary==2.9.9
requests==2.31.0
python-multipart==0.0.6



bcrypt==4.1.2
PyJWT==2.8.0
python-jose[cryptography]==3.3.0